Make the github notifications accessible.
Make the github notifications accessible. this adds support by making every repo an h2 instead of an h3, so that each notification can be an h3. Further, it moves the issue graphic after the link, so the link is the first thing that comes up as a notification. A new helper "makeElementOwn" is added to make aria-ownsing something super easy.