Google Keep: Fix moving focus when search is dismissed.
Previously, this code was triggering when it shouldn't such as when editing a new note, causing focus to get spuriously lost.
Now, we use the visibility of the Clear search button to figure out whether search was dismissed.
This required adding a whenAttrChangedOnAncestor option to prevent a tweak from being applied when an attribute changes on an ancestor.
Otherwise, every time the style attribute changed on an ancestor, this code would trigger and throw focus.